IframeLabelAlternative.tsx renamed to FieldLabelAlternative.tsx now it is not just used for securedField related inputs
