+++ packages/lib/src/components/Ach/components/AchInput/components/AchSFInput.tsx
@@ -2,7 +2,7 @@ import { h } from 'preact';
 import classNames from 'classnames';
 import Field from '../../../../internal/FormFields/Field';
 import DataSfSpan from '../../../../Card/components/CardInput/components/DataSfSpan';
-import { alternativeLabelContent } from '../../../../Card/components/CardInput/components/IframeLabelAlternative';
+import { alternativeLabelContent } from '../../../../Card/components/CardInput/components/FieldLabelAlternative';
 import { capitalizeFirstLetter } from '../../../../../utils/textUtils';
 
 const AchSFInput = ({ id, dataInfo, className = '', label, focused, filled, errorMessage = '', isValid = false, onFocusField, dir }) => {
