+++ packages/lib/src/components/Giftcard/components/GiftcardNumberField.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import Field from '../../internal/FormFields/Field';
 import { h } from 'preact';
 import { GiftcardFieldProps } from './types';
-import { alternativeLabelContent } from '../../Card/components/CardInput/components/IframeLabelAlternative';
+import { alternativeLabelContent } from '../../Card/components/CardInput/components/FieldLabelAlternative';
 
 export const GiftcardNumberField = ({ i18n, classNameModifiers, sfpState, getCardErrorMessage, focusedElement, setFocusOn }: GiftcardFieldProps) => {
     return (
